manage test-resources
